requirements.txt
jupyter
pandas==1.5.3
matplotlib==3.6.3
torch==1.13.1
pytorch-lightning==1.9.0
transformers==4.26.0
bitsandbytes-cuda117==0.26.0.post2
accelerate==0.16.0
wandb
openai